backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
python-dotenv
httpx
pgvector
sentence-transformers
transformers
torch
langchain
langchain-core
langchain-community
langchain-ollama
langchain-groq
motor
pymongo
